Split the assistant panel, and stop throwing away work that is still running
rocprofvis_ai_assistant.cpp carried three jobs at 1783 lines. The standing instructions are content rather than code and now live in rocprofvis_ai_prompts.cpp, where they can be read and revised as prose; the ImGui half moves to rocprofvis_ai_assistant_render.cpp, which touches no conversation state. What remains is the turn machine, at 945.

A tool called while a trace was still opening returned 'still loading' in microseconds, and nothing made the model wait - so on a large file every tool in every round failed instantly and the whole round budget burned before the trace was readable. That wait now parks and polls like any other fetch.

The fetch deadline moves to 350s with the elapsed time shown in the status, because cutting a live query off and reporting a timeout is worse than making the user wait: the model answers around the gap instead of reporting it.

TrimConversation could not bound one investigation. It only cut on a user message, and a dive is a single question followed by many rounds of tool replies, so it never fired while the transcript grew past anything the endpoint would accept. Old tool replies are now compacted in place, which keeps each tool_call_id paired with the call that asked for it.

Tool arguments and results are logged at debug, so an answer can be checked against what the model was actually given. That is what found the table decoding bug.
